Add files via upload
OIBSIP
Task 2- Online Examination
This system is made by using java (AWT), MySQL, and SQL queries. In this system, first user registered by name, email,  username, and password and this information is save into the database. Then the user can log in, if anybody wants to register with the same mail id then it can't be possible. After this user has to enter the correct username and password to log in. After this user can also update the password by entering the correct current password and that password will also save in the database. Now user can start the test in a given time and after clicking submit button score will be displayed and the exit button is there to close the window.
I use eclipse ide for this MySQL. By using MySQL I create a database named "name" and create a table having columns "name", "emailid", "username", "password".
Thank you.....